Expose pusher executable attribute for users
This lets people get around the limitation of using the pusher from the
python containerregistry, and can substitute with their own pusher.

The main issue with the python containerregistry is that it uses python2
httplib2, and so workspaces using python3 as their interpreter can not
depend on it successfully.

See for instance, a pusher in go at
bazelbuild#481.